Split src/utils/utils.ts
