+++ src/providers/preview/mathpreview.ts
@@ -2,7 +2,7 @@ import * as vscode from 'vscode'
 import type {TypesetArg} from 'mathjax-node'
 
 import {MathJaxPool} from './mathjaxpool'
-import * as utils from '../../utils/utils'
+import * as utils from '../../utils/svg'
 import type {Extension} from '../../main'
 import type {ReferenceEntry} from '../completer/reference'
 import {getCurrentThemeLightness} from '../../utils/theme'
