+++ src/providers/preview/mathpreviewlib/hoverpreviewonref.ts
@@ -1,6 +1,6 @@
 import * as vscode from 'vscode'
 import type {TypesetArg} from 'mathjax-node'
-import * as utils from '../../../utils/utils'
+import * as utils from '../../../utils/svg'
 import type {MathJaxPool} from '../mathjaxpool'
 import type {ReferenceEntry} from '../../completer/reference'
 import type {Extension} from '../../../main'
