+++ src/utils/svg.ts
@@ -1,0 +1,8 @@
+
+export function svgToDataUrl(xml: string): string {
+    // We have to call encodeURIComponent and unescape because SVG can includes non-ASCII characters.
+    // We have to encode them before converting them to base64.
+    const svg64 = Buffer.from(unescape(encodeURIComponent(xml)), 'binary').toString('base64')
+    const b64Start = 'data:image/svg+xml;base64,'
+    return b64Start + svg64
+}
